# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Refactoring",
        "title": "OnlineJudge",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Eutop1a",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:65533",
    "basePath": "/api/v1",
    "paths": {
        "/api/sort": {
            "post": {
                "description": "获取所有用户并按分数降序排序",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "根据用户分数降序排序",
                "parameters": [
                    {
                        "type": "string",
                        "description": "执行操作的用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取并排序用户列表",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseSort"
                        }
                    },
                    "403": {
                        "description": "Token 已超时",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "数据库查询错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            }
        },
        "/changeScore": {
            "post": {
                "description": "通过用户名和新分数的表单数据，增加用户的总分数，并返回新的总分数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "增加用户分数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "要增加的分数",
                        "name": "newscore",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "增加分数成功，返回消息和新的总分数",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseAddScore"
                        }
                    },
                    "403": {
                        "description": "Token 已超时",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "数据库查询或保存出错",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            }
        },
...
        "/list": {
            "post": {
                "description": "获取所有问题的列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取问题列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "要获取问题列表的用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取问题列表",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseProblems"
                        }
                    },
                    "403": {
                        "description": "Token 已超时",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "获取问题列表出错",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseMsg"
                        }
                    }
                }
            }
        },
...
        "/problem/:id": {
            "post": {
                "description": "通过问题ID和管理员身份，获取指定问题的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取问题详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "执行操作的管理员用户名",
                        "name": "_",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "要获取的问题ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取问题详情",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseQuestionDetail"
                        }
                    },
                    "403": {
                        "description": "获取问题详情出错",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseMsg"
                        }
                    }
                }
            }
        },
        "/problem/add": {
            "post": {
                "description": "通过管理员身份，使用表单数据添加新的问题，并返回操作结果",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "添加问题",
                "parameters": [
                    {
                        "type": "string",
                        "description": "执行操作的管理员用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "问题ID",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "问题标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "问题描述",
                        "name": "lore",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "问题标准输入",
                        "name": "input",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "问题标准输出",
                        "name": "output",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "问题提示",
                        "name": "tips",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功添加问题",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseAddProblems"
                        }
                    },
                    "403": {
                        "description": "解析表单数据出错",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseMsg"
                        }
                    }
                }
            }
        },
        "/problem/file/:id": {
            "post": {
                "description": "通过问题ID，获取指定问题的文件列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取问题文件列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "要获取文件列表的问题ID",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取问题文件列表",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseGetData"
                        }
                    },
                    "403": {
                        "description": "读取文件列表出错",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseDataString"
                        }
                    }
                }
            }
        },
        "/problem/file/add/:id": {
            "post": {
                "description": "通过管理员身份，使用表单数据上传问题的输入和输出文件，并返回操作结果",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "上传问题文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "执行操作的管理员用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "要上传文件的问题ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "问题输入文件（.in）",
                        "name": "input",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "问题输出文件（.out）",
                        "name": "output",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功上传问题文件",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "接收或保存文件出错",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseMsg"
                        }
                    },
                    "403": {
                        "description": "Token 已超时或用户非管理员",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseMsg"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller._ResponseAddProblems": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
        "controller._ResponseAddScore": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "new_score": {
                    "type": "integer"
                }
            }
        },
        "controller._ResponseDataString": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                }
            }
        },
        "controller._ResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "controller._ResponseGetData": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "controller._ResponseMsg": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "controller._ResponseProblems": {
            "type": "object"
        },
        "controller._ResponseQuestionDetail": {
            "type": "object"
        },
        "controller._ResponseSort": {
            "type": "object"
        },
...
    }
}